+++ src/components/SwitchMusic/SpotifySwitch/UserSpotify/UserPlaylist.js
@@ -2,47 +2,10 @@ import React, { useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
 import queryString from 'query-string';
 import axios from 'axios';
+import UserContainer from '@components/SwitchMusic/SpotifySwitch/UserSpotify/User.container/';
 
-const UserPlaylist = ({ location }) => {
-  // Token values
-  const [accesToken, setAccessToken] = useState('');
-  const [tokenType, setTokenType] = useState('');
-  const [expiresIn, setExpiresIn] = useState('');
-  const [spotifyPlaylistData, setSpotifyPlaylistData] = useState({});
-
-  // Hash Token
-  const { hash } = location;
-  const hashValues = queryString.parse(hash);
-  const { access_token, expires_in, token_type } = hashValues;
-
+const UserPlaylist = ({ spotifyPlaylistData }) => {
   const { items } = spotifyPlaylistData;
-
-  // Token headers for the get request
-  const config = {
-    headers: { 'Content-Type': 'application/json', Authorization: `Bearer ${access_token}` },
-  };
-
-  useEffect(() => {
-    setAccessToken(access_token);
-    setTokenType(expires_in);
-    setExpiresIn(token_type);
-  }, []);
-
-  useEffect(() => {
-    async function fetchData() {
-      if (accesToken) {
-        try {
-          const result = await axios.get(SPOTIFY_PROFIL_PLAYLISTS, config);
-          setSpotifyPlaylistData(result.data);
-        } catch (err) {
-          console.error('Error to fetch Spotify profile data');
-        }
-      }
-    }
-    fetchData();
-  }, [accesToken, SPOTIFY_PROFIL_PLAYLISTS]);
-
-  console.log(items);
 
   return (
     <div>
@@ -62,4 +25,4 @@ const UserPlaylist = ({ location }) => {
   );
 };
 
-export default UserPlaylist;
+export default UserContainer(UserPlaylist);
