+++ src/components/SwitchMusic/SpotifySwitch/UserSpotify/UserProfile.js
@@ -1,45 +1,11 @@
 import React, { useEffect, useState } from 'react';
 import queryString from 'query-string';
 import axios from 'axios';
+import UserContainer from '@components/SwitchMusic/SpotifySwitch/UserSpotify/User.container/';
 
-const UserProfile = ({ location }) => {
-  // Token values
-  const [accesToken, setAccessToken] = useState('');
-  const [tokenType, setTokenType] = useState('');
-  const [expiresIn, setExpiresIn] = useState('');
-  const [spotifyProfilData, setSpotifyProfilData] = useState({});
-
-  // Hash Token
-  const { hash } = location;
-  const hashValues = queryString.parse(hash);
-  const { access_token, expires_in, token_type } = hashValues;
+const UserProfile = ({ spotifyProfilData }) => {
 
   const { country, display_name, product, followers, external_urls, images } = spotifyProfilData;
-
-  // Token headers for the get request
-  const config = {
-    headers: { 'Content-Type': 'application/json', Authorization: `Bearer ${access_token}` },
-  };
-
-  useEffect(() => {
-    setAccessToken(access_token);
-    setTokenType(expires_in);
-    setExpiresIn(token_type);
-  }, []);
-
-  useEffect(() => {
-    async function fetchData() {
-      if (accesToken) {
-        try {
-          const result = await axios.get(SPOTIFY_PROFIL, config);
-          setSpotifyProfilData(result.data);
-        } catch (err) {
-          console.error('Error to fetch Spotify profile data');
-        }
-      }
-    }
-    fetchData();
-  }, [accesToken, SPOTIFY_PROFIL]);
 
   return (
     <div>
@@ -60,4 +26,4 @@ const UserProfile = ({ location }) => {
   );
 };
 
-export default UserProfile;
+export default UserContainer(UserProfile);
